task8/requirements.txt
apache-airflow
apache-airflow-providers-snowflake
pyarrow==10.0.1
pandas